+++ tobago-core/src/main/java/org/apache/myfaces/tobago/internal/component/AbstractUIButtons.java
@@ -19,12 +19,28 @@
 
 package org.apache.myfaces.tobago.internal.component;
 
+import org.apache.myfaces.tobago.component.SupportsDisabledProperty;
 import org.apache.myfaces.tobago.layout.Orientation;
 
 /**
  * {@link org.apache.myfaces.tobago.internal.taglib.component.ButtonsTagDeclaration}
  */
-public abstract class AbstractUIButtons extends AbstractUIPanelBase {
+public abstract class AbstractUIButtons extends AbstractUIPanelBase implements SupportsDisabledProperty {
+
+  enum PropertyKeys {
+    disabled,
+  }
 
   public abstract Orientation getOrientation();
+
+  @Override
+  public boolean isDisabled() {
+    final Boolean disabled = (Boolean) getStateHelper().eval(PropertyKeys.disabled);
+    return disabled != null && disabled;
+  }
+
+  @Override
+  public void setDisabled(final boolean disabled) {
+    getStateHelper().put(PropertyKeys.disabled, disabled);
+  }
 }
