+++ tobago-example/tobago-example-demo/src/main/webapp/content/20-component/040-command/20-buttons/Button_Group.xhtml
@@ -94,4 +94,19 @@
       <tc:button label="Right"/>
     </tc:buttons>
   </tc:section>
+
+  <tc:section label="Disable">
+    <p>This example shows how to disable all buttons in a button group.
+      It's also possible to set disable to false e.g. for one button.</p>
+    <pre><code class="language-markup">&lt;tc:buttons disabled="true">
+  &lt;c:button label="Left" disabled="false"/>
+  &lt;tc:button label="Center"/>
+  &lt;tc:button label="Right"/>
+&lt;/tc:buttons></code></pre>
+    <tc:buttons disabled="true">
+      <tc:button label="Left" disabled="false"/>
+      <tc:button label="Center"/>
+      <tc:button label="Right"/>
+    </tc:buttons>
+  </tc:section>
 </ui:composition>
